% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/couple.R
\name{apply_cca_prediction}
\alias{apply_cca_prediction}
\title{Apply CCA Prediction Transform}
\usage{
apply_cca_prediction(new_amplitudes, cca_result, k)
}
\arguments{
\item{new_amplitudes}{Tibble with time and predictor PC amplitudes}

\item{cca_result}{CCA result object from cancor()}

\item{k}{Number of CCA modes to use}
}
\value{
Tibble with time and predicted response PC amplitudes
}
\description{
Internal function that applies the CCA transformation to predict response amplitudes
from predictor amplitudes.
}
\keyword{internal}
