% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/patterns_class.R
\name{as.data.frame.patterns}
\alias{as.data.frame.patterns}
\title{Format patterns as a data frame}
\usage{
\method{as.data.frame}{patterns}(x, ...)
}
\arguments{
\item{x}{A patterns object}

\item{...}{Additional arguments}
}
\description{
Format patterns as a data frame
}
