% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/metrics.R
\name{calc_rmse}
\alias{calc_rmse}
\title{Calculate root mean squared error}
\usage{
calc_rmse(pred_matrix, obs_matrix)
}
\arguments{
\item{pred_matrix}{Predicted values matrix (time x space)}

\item{obs_matrix}{Observed values matrix (time x space)}
}
\value{
Scalar RMSE value
}
\description{
Calculate root mean squared error
}
\keyword{internal}
