% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/patterns_class.R
\name{$.patterns}
\alias{$.patterns}
\title{Extract components from patterns object}
\usage{
\method{$}{patterns}(x, name)
}
\arguments{
\item{x}{A patterns object}

\item{name}{Name of component to extract}
}
\value{
The requested component
}
\description{
Extract components from patterns object
}
