% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/couple.R
\name{check_cca_method}
\alias{check_cca_method}
\title{Guard a CCA-only accessor against non-CCA coupled objects}
\usage{
check_cca_method(object, fn, call = rlang::caller_env())
}
\description{
Guard a CCA-only accessor against non-CCA coupled objects
}
\keyword{internal}
