% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/patterns_core.R
\name{check_k_valid}
\alias{check_k_valid}
\title{Check if k is valid}
\usage{
check_k_valid(k, max_k, arg = rlang::caller_arg(k), call = rlang::caller_env())
}
\arguments{
\item{k}{Number of components}

\item{max_k}{Maximum allowed components}

\item{arg}{Argument name for error messages}

\item{call}{Calling environment for error messages}
}
\description{
Check if k is valid
}
\keyword{internal}
