% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/patterns_core.R
\name{check_single_attribute}
\alias{check_single_attribute}
\title{Check that a stars object has exactly one attribute}
\usage{
check_single_attribute(
  x,
  arg = rlang::caller_arg(x),
  call = rlang::caller_env()
)
}
\arguments{
\item{x}{A stars object}

\item{arg}{Argument name for error messages}

\item{call}{Calling environment for error messages}
}
\description{
EOF analysis operates on a single variable; silently using the first
attribute of a multi-attribute object would hide the others.
}
\keyword{internal}
