% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/patterns_core.R
\name{check_stars_object}
\alias{check_stars_object}
\title{Check if input is a stars object}
\usage{
check_stars_object(x, arg = rlang::caller_arg(x), call = rlang::caller_env())
}
\arguments{
\item{x}{Object to check}

\item{arg}{Argument name for error messages}

\item{call}{Calling environment for error messages}
}
\description{
Check if input is a stars object
}
\keyword{internal}
