% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/delta.R
\name{delta_mul}
\alias{delta_mul}
\title{Delta change methods}
\usage{
delta_mul(pred, obs, newdata = NULL, k = NULL)
}
\arguments{
\item{k}{A placeholder, must be NULL.}
}
\value{

}
\description{
Delta change methods
}
