% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stars_utils.R
\name{dimension_size}
\alias{dimension_size}
\title{Obtain the size of a stars dimension definition}
\usage{
dimension_size(dimension)
}
\description{
Obtain the size of a stars dimension definition
}
\keyword{internal}
