% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_patterns.R
\name{.plot_amplitudes_internal}
\alias{.plot_amplitudes_internal}
\title{Internal function for amplitude plotting}
\usage{
.plot_amplitudes_internal(
  x,
  scale = "standardized",
  scale_y = "fixed",
  events = NULL,
  layout = NULL
)
}
\description{
Internal function for amplitude plotting
}
\keyword{internal}
