% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_patterns.R
\name{.simple_smart_layout}
\alias{.simple_smart_layout}
\title{Simple smart layout function - KISS principle}
\usage{
.simple_smart_layout(k)
}
\description{
Simple smart layout function - KISS principle
}
\keyword{internal}
