% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tune_cca.R
\name{evaluate_fold}
\alias{evaluate_fold}
\title{Evaluate a single fold with given parameters}
\usage{
evaluate_fold(fold, k_pred, k_resp, k_cca, method = "cca", metrics)
}
\arguments{
\item{fold}{A single fold list from cv_folds$folds}

\item{k_pred}{Number of predictor EOFs}

\item{k_resp}{Number of response EOFs}

\item{k_cca}{Number of CCA modes}

\item{method}{Coupling method passed to couple() ("cca" or "pcr")}

\item{metrics}{Metrics to compute}
}
\value{
Tibble with fold_id and metric values
}
\description{
Evaluate a single fold with given parameters
}
\keyword{internal}
