% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/legacy_wrappers.R
\name{fit_model}
\alias{fit_model}
\title{Legacy GAM Model Fitting (Deprecated)}
\usage{
fit_model(data_in)
}
\arguments{
\item{data_in}{Prepared data from prep_data}
}
\value{
Fitted GAM models
}
\description{
Legacy GAM Model Fitting (Deprecated)
}
