% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/couple.R
\name{get_canonical_variables}
\alias{get_canonical_variables}
\title{Get Canonical Variables from Coupled Patterns}
\usage{
get_canonical_variables(
  object,
  data,
  type = c("predictor", "response"),
  k = NULL
)
}
\arguments{
\item{object}{A coupled_patterns object}

\item{data}{Original data (patterns object or amplitudes tibble)}

\item{type}{Either "predictor" or "response"}

\item{k}{Number of canonical modes to extract}
}
\value{
Tibble with canonical variables
}
\description{
Extract canonical variables from either predictor or response patterns
}
