% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/teleconnections.R
\name{get_corrs}
\alias{get_corrs}
\title{Calculate teleconnections between a patterns object and another layer}
\usage{
get_corrs(dat, patterns)
}
\arguments{
\item{patterns}{}
}
\value{

}
\description{
Calculate teleconnections between a patterns object and another layer
}
