% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_patterns.R
\name{get_patterns}
\alias{get_patterns}
\title{Get EOFs and PCs from observations}
\usage{
get_patterns(
  dat,
  k = 4,
  scale = FALSE,
  rotate = FALSE,
  monthly = FALSE,
  weight = TRUE
)
}
\arguments{
\item{dat}{A `stars` object containing spatial and temporal dimensions.}

\item{k}{The number of PC/EOF modes to retain.}
}
\value{
A `patterns` object containing a tibble of PC amplitudes, a `stars`
object with EOF spatial patterns, and a `stars` object with the mean and
(optionally) standard deviation fields used.
}
\description{
Get EOFs and PCs from observations
}
