% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_pcs.R
\name{get_pcs}
\alias{get_pcs}
\title{Calculate principal components}
\usage{
get_pcs(dat, time_dim = 3, scale = FALSE, clim = NULL)
}
\arguments{
\item{scale}{}
}
\value{

}
\description{
Calculate principal components
}
