% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/patterns_core.R
\name{get_spatial_dimensions}
\alias{get_spatial_dimensions}
\title{Get spatial dimension names for stars object}
\usage{
get_spatial_dimensions(dat)
}
\arguments{
\item{dat}{A stars object}
}
\value{
Character vector of spatial dimension names
}
\description{
Get spatial dimension names for stars object
}
\keyword{internal}
