% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/patterns_class.R
\name{get_times}
\alias{get_times}
\title{Extract time vector from a patterns object or data frame}
\usage{
get_times(x)
}
\arguments{
\item{x}{A patterns object or a data.frame/tibble with a time column}
}
\value{
A vector of time values
}
\description{
Extract time vector from a patterns object or data frame
}
\keyword{internal}
