% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/patterns_core.R
\name{has_geometry_dimension}
\alias{has_geometry_dimension}
\title{Check if stars object has sf geometry}
\usage{
has_geometry_dimension(dat)
}
\arguments{
\item{dat}{A stars object}
}
\value{
Logical indicating if object has geometry dimension
}
\description{
Check if stars object has sf geometry
}
\keyword{internal}
