% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/teleconnections.R
\name{match_times}
\alias{match_times}
\title{Match time steps between amplitudes and a stars object}
\usage{
match_times(amplitudes, dat)
}
\arguments{
\item{amplitudes}{A tibble with a time column}

\item{dat}{A stars object with a time dimension}
}
\value{
A list with `times` (common time steps) and `amps` (filtered amplitude matrix)
}
\description{
Match time steps between amplitudes and a stars object
}
\keyword{internal}
