% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_climatology.R
\name{month_dimension}
\alias{month_dimension}
\title{Build a calendar month dimension (values 1:12)}
\usage{
month_dimension()
}
\description{
Build a calendar month dimension (values 1:12)
}
\keyword{internal}
