% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_patterns.R
\name{plot_scree}
\alias{plot_scree}
\title{Scree plot (deprecated)}
\usage{
plot_scree(
  dat,
  k = NULL,
  kmax = 10,
  scale = FALSE,
  monthly = FALSE,
  weight = TRUE
)
}
\arguments{
\item{dat}{Data object}

\item{k}{Number of components}

\item{kmax}{Maximum components to show}

\item{scale}{Scaling option}

\item{monthly}{Monthly option}

\item{weight}{Weighting option}
}
\description{
Scree plot (deprecated)
}
\details{
\strong{This function is deprecated.} Use \code{screeplot(patterns)} instead
where patterns is the result of \code{get_patterns()}.
}
