% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/legacy_wrappers.R
\name{predict_eot}
\alias{predict_eot}
\title{Legacy EOT Prediction (Deprecated)}
\usage{
predict_eot(cv, k)
}
\arguments{
\item{cv}{Cross-validation object from prep_eot}

\item{k}{Number of modes}
}
\value{
Reconstructed field
}
\description{
Legacy EOT Prediction (Deprecated)
}
