% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prep_cv.R
\name{prep_delta}
\alias{prep_delta}
\title{Prepare delta method cross-validation data}
\usage{
prep_delta(preds, obs)
}
\arguments{
\item{preds}{Predictor stars object}

\item{obs}{Observed/response stars object}
}
\value{
Tibble with train_preds, train_obs, and test columns for each fold
}
\description{
Sets up k-fold cross-validation for delta-based downscaling.
}
