% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/preprocess.R
\name{preprocess}
\alias{preprocess}
\title{Preprocessing functions}
\usage{
preprocess(x, var, flip = FALSE, regrid = FALSE, monthly = FALSE)
}
\arguments{
\item{monthly}{}
}
\value{

}
\description{
Preprocessing functions
}
