% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/couple.R
\name{print.coupled_patterns}
\alias{print.coupled_patterns}
\title{Print method for coupled_patterns}
\usage{
\method{print}{coupled_patterns}(x, ...)
}
\arguments{
\item{x}{A coupled_patterns object}

\item{...}{Additional arguments (ignored)}
}
\description{
Print method for coupled_patterns
}
