% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tune_cca.R
\name{print.cv_folds}
\alias{print.cv_folds}
\title{Print method for cv_folds objects}
\usage{
\method{print}{cv_folds}(x, ...)
}
\arguments{
\item{x}{A cv_folds object}

\item{...}{Additional arguments (ignored)}
}
\description{
Print method for cv_folds objects
}
