% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/patterns_class.R
\name{print.patterns}
\alias{print.patterns}
\title{Print method for patterns objects}
\usage{
\method{print}{patterns}(x, ...)
}
\arguments{
\item{x}{A patterns object}

\item{...}{Additional arguments passed to print}
}
\description{
Print method for patterns objects
}
