% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/patterns_class.R
\name{print.summary.patterns}
\alias{print.summary.patterns}
\title{Print method for summary.patterns objects}
\usage{
\method{print}{summary.patterns}(x, ...)
}
\arguments{
\item{x}{A summary.patterns object}

\item{...}{Additional arguments passed to print}
}
\description{
Print method for summary.patterns objects
}
