% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_patterns.R
\name{print_significance}
\alias{print_significance}
\title{Print significance test results}
\usage{
print_significance(signif_obj)
}
\arguments{
\item{signif_obj}{Significance object from calculate_significance}
}
\description{
Print significance test results
}
