% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/legacy_wrappers.R
\name{show_migration_guide}
\alias{show_migration_guide}
\title{Migration Helper: Show How to Update Code}
\usage{
show_migration_guide(
  legacy_function = c("predict_cca", "predict_gam", "predict_pcr", "predict_eot",
    "couple_patterns")
)
}
\arguments{
\item{legacy_function}{Name of the legacy function you were using}
}
\description{
This function provides guidance on how to migrate from legacy functions
to the new CCA-based architecture.
}
