% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/patterns_class.R
\name{[.patterns}
\alias{[.patterns}
\title{Extract subset of PCs from a patterns object}
\usage{
\method{[}{patterns}(x, i, ...)
}
\arguments{
\item{x}{A patterns object}

\item{i}{Index or names of PCs to keep}

\item{...}{Additional arguments (unused)}
}
\value{
A patterns object with only selected PCs
}
\description{
Extract subset of PCs from a patterns object
}
