% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/couple_patterns.R
\name{summary.coupled_patterns}
\alias{summary.coupled_patterns}
\title{Summary method for coupled_patterns}
\usage{
\method{summary}{coupled_patterns}(object, ...)
}
\arguments{
\item{object}{A coupled_patterns object}

\item{...}{Additional arguments (ignored)}
}
\description{
Summary method for coupled_patterns
}
