% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/patterns_class.R
\name{summary.patterns}
\alias{summary.patterns}
\title{Summary method for patterns objects}
\usage{
\method{summary}{patterns}(object, ...)
}
\arguments{
\item{object}{A patterns object}

\item{...}{Additional arguments passed to summary}
}
\description{
Summary method for patterns objects
}
