% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_patterns.R
\name{.plot_eofs_internal}
\alias{.plot_eofs_internal}
\title{Internal function for EOF plotting}
\usage{
.plot_eofs_internal(
  x,
  scaled = FALSE,
  rawdata = NULL,
  layout = NULL,
  overlay = NULL,
  overlay_color = "grey30",
  overlay_fill = NA
)
}
\description{
Internal function for EOF plotting
}
\keyword{internal}
