% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/legacy_wrappers.R
\name{fit_pcr}
\alias{fit_pcr}
\title{Legacy PCR Model Fitting (Deprecated)}
\usage{
fit_pcr(data_in)
}
\arguments{
\item{data_in}{Prepared data from prep_data}
}
\value{
Fitted PCR models
}
\description{
Legacy PCR Model Fitting (Deprecated)
}
