% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_climatology.R
\name{get_climatology}
\alias{get_climatology}
\title{Get annual or monthly means and standard deviations, or rescale a grid using climatology}
\usage{
get_climatology(dat, monthly = FALSE)
}
\arguments{
\item{dat}{Input data.}

\item{monthly}{Calculate monthly climatology or anomalies instead of using
the entire period? Defaults to FALSE.}

\item{clim}{An optional climatology made with get_climatology().
Defaults to NULL, in which case get_climatology() is called internally.}

\item{scale}{Calculate standardized anomalies (x - mean) / sd? Defaults to
FALSE.

Based on the `climatology()` and `scaleGrid()` functions from transformeR.
These functions apply calculations per grid cell.}
}
\description{
Get annual or monthly means and standard deviations, or rescale a grid using climatology
}
