% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/teleconnections.R
\name{get_correlation}
\alias{get_correlation}
\title{Calculate teleconnections between a patterns object and another layer}
\usage{
get_correlation(dat, patterns, amplitudes = NULL)
}
\arguments{
\item{patterns}{}
}
\description{
The function will check if there are any overlapping time steps between the
raster field and the PC amplitudes.
}
