% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_eofs.R
\name{get_eofs}
\alias{get_eofs}
\title{Title}
\usage{
get_eofs(dat, pca, k, rotate = FALSE)
}
\arguments{
\item{rotate}{scale by stdev (i.e. sqrt(eigenvalues)) for more robust rotation (Hannachi et al 2007)}
}
\description{
Title
}
