% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_patterns.R
\name{get_patterns}
\alias{get_patterns}
\title{Get EOFs and PCs from observations}
\usage{
get_patterns(dat, k = 4, scale = FALSE, rotate = FALSE)
}
\arguments{
\item{k}{}
}
\value{

}
\description{
Get EOFs and PCs from observations
}
\examples{

}
