% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_pcs.R
\name{get_pcs}
\alias{get_pcs}
\title{Calculate principal components}
\usage{
get_pcs(dat, scale = FALSE, clim = NULL, monthly = FALSE, weight = TRUE)
}
\arguments{
\item{scale}{}
}
\description{
Calculate principal components
}
