% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_patterns.R
\name{plot_eofs}
\alias{plot_eofs}
\title{Plotting functions for patterns objects}
\usage{
plot_eofs(patterns, scaled = FALSE, rawdata = NULL)
}
\arguments{
\item{normalized}{}
}
\description{
Plotting functions for patterns objects
}
