% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reconstruct_field.R
\name{should_be_nonnegative}
\alias{should_be_nonnegative}
\title{Title}
\usage{
should_be_nonnegative(patterns)
}
\arguments{
\item{patterns}{A patterns object}

\item{amplitudes}{}
}
\value{
Logical indicating if non-negative constraint should apply
}
\description{
Title
}
\examples{
Detect if variable should be non-negative
}
\keyword{internal}
